requirements.txt
streamlit
matplotlib
fastapi
uvicorn
pymssql==2.3.1
python-dotenv
firebirdsql==1.3.1
passlib==1.7.4